Add basic test
